+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/multiblocks/miner/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package holds classes associated with the
+ * {@link io.github.thebusybiscuit.slimefun4.implementation.items.multiblocks.IndustrialMiner}.
+ */
+package io.github.thebusybiscuit.slimefun4.implementation.items.multiblocks.miner;
